+++ packages/perseus/src/widgets/interactive-graphs/graphs/linear.tsx
@@ -1,10 +1,10 @@
 import * as React from "react";
 
 import {usePerseusI18n} from "../../../components/i18n-context";
-import a11y from "../../../util/a11y";
 import {actions} from "../reducer/interactive-graph-action";
 
 import {MovableLine} from "./components/movable-line";
+import SRDescInSVG from "./components/sr-description-within-svg";
 import {srFormatNumber} from "./screenreader-text";
 import {getInterceptStringForLine, getSlopeStringForLine} from "./utils";
 
@@ -81,15 +81,13 @@ const LinearGraph = (props: LinearGraphProps, key: number) => {
             />
             {/* Hidden elements to provide the descriptions for the
                 circle and radius point's `aria-describedby` properties. */}
-            <g id={pointsDescriptionId} style={a11y.srOnly}>
+            <SRDescInSVG id={pointsDescriptionId}>
                 {srLinearGraphPoints}
-            </g>
-            <g id={interceptDescriptionId} style={a11y.srOnly}>
+            </SRDescInSVG>
+            <SRDescInSVG id={interceptDescriptionId}>
                 {interceptString}
-            </g>
-            <g id={slopeDescriptionId} style={a11y.srOnly}>
-                {slopeString}
-            </g>
+            </SRDescInSVG>
+            <SRDescInSVG id={slopeDescriptionId}>{slopeString}</SRDescInSVG>
         </g>
     );
 };